Changes to using agpypeline library
